simplifying transfer views
