--- project_electron/orgs/templates/orgs/recent_transfers.html
@@ -1,38 +1,0 @@
-{% extends 'transfer_app/base.html' %}
-{% load tz %}
-
-{% block h1_title %}Recent Transfers{% endblock %}
-
-{% block content %}
-<div class="row">
-	<div class="col-md-12">
-		<div class="nav-tabs-custom">
-			<ul class="nav nav-tabs">
-				<li class="active"><a href="#tab_org" data-toggle="tab">{% if request.user.is_archivist %}All{% else %}{{ request.user.organization }}{% endif %} Transfers ({{ org_uploads_count }})</a></li>
-				<li class=""><a href="#tab_user" data-toggle="tab">My Transfers ({{ user_uploads_count }})</a></li>
-			</ul>
-			<div class="tab-content">
-				<div class="tab-pane active table-responsive" id="tab_org">
-
-					{% include 'parts/transfer_table.html' with org=request.user.organization uploads=org_uploads %}
-
-					<div class="mar-top20 text-center">
-						<a href="{% url 'orgs-transfers-report' pk=request.user.organization.pk %}" class="btn btn-info">See all {% if not request.user.is_archivist %}{{ request.user.organization }} {% endif %}transfers</a>
-					</div>
-
-				</div>
-				<div class="tab-pane table-responsive" id="tab_user">
-
-					{% include 'parts/transfer_table.html' with org=request.user.organization uploads=user_uploads %}
-
-					<div class="mar-top20 text-center">
-						<a href="{% url 'users-transfers-report' pk=request.user.pk %}" class="btn btn-info">See all my transfers</a>
-					</div>
-
-				</div>
-			</div>
-		</div>
-	</div>
-</div>
-
-{% endblock %}
