+++ project_electron/orgs/templates/orgs/transfers.html
@@ -1,0 +1,59 @@
+{% extends 'transfer_app/base.html' %}
+{% load tz %}
+
+{% block h1_title %}
+Transfers
+<a href="{% url 'transfer-data' %}" class="btn btn-primary pull-right"><i class="fa fa-download"></i> Download CSV</a>
+{% endblock %}
+
+{% block content %}
+<div class="row">
+	<div class="col-md-12">
+		<div class="nav-tabs-custom">
+			<ul class="nav nav-tabs">
+				<li class="active"><a href="#tab_org" data-toggle="tab">{% if request.user.is_archivist %}All{% else %}{{ request.user.organization }}{% endif %} Transfers ({{ org_uploads_count }})</a></li>
+				<li class=""><a href="#tab_user" data-toggle="tab">My Transfers ({{ user_uploads_count }})</a></li>
+			</ul>
+			<div class="tab-content">
+				<div class="tab-pane active table-responsive" id="tab_org">
+
+					{% include 'parts/transfer_table.html' with uploads=org_uploads table_id='org_transfers' %}
+
+				</div>
+				<div class="tab-pane table-responsive" id="tab_user">
+
+					{% include 'parts/transfer_table.html' with uploads=user_uploads table_id='user_transfers' %}
+
+				</div>
+			</div>
+		</div>
+	</div>
+</div>
+
+{% endblock %}
+
+{% block extra_js %}
+<script type="text/javascript">
+  $(document).ready(function() {
+    $('#org_transfers').DataTable({
+      'paging'      : true,
+      'lengthChange': false,
+      'searching'   : true,
+      'ordering'    : true,
+      'info'        : true,
+      'autoWidth'   : true,
+      'pageLength'  : 25
+    })
+		$('#user_transfers').DataTable({
+      'paging'      : true,
+      'lengthChange': false,
+      'searching'   : true,
+      'ordering'    : true,
+      'info'        : true,
+      'autoWidth'   : true,
+      'pageLength'  : 25
+    })
+    $.fn.dataTable.moment( 'MMM. D, YYYY, h:m a' );
+  })
+</script>
+{% endblock %}
