+++ project_electron/orgs/urls.py
@@ -7,6 +7,4 @@ urlpatterns = [
     url(r'^(?P<pk>\d+)/$', OrganizationDetailView.as_view(), name='orgs-detail'),
     url(r'^$', 	OrganizationListView.as_view(), name='orgs-list'),
     url(r'^(?P<pk>\d+)/edit/$', OrganizationEditView.as_view(), name='orgs-edit'),
-    url(r'^(?P<pk>\d+)/transfers/$', OrganizationTransfersView.as_view(), name='orgs-transfers-report'),
-    url(r'^(?P<pk>\d+)/transfers/csv/$', OrganizationTransferDataView.as_view(), name='orgs-transfer-data'),
 ]
