+++ project_electron/orgs/user_urls.py
@@ -7,7 +7,6 @@ urlpatterns = [
     url(r'^(?P<pk>\d+)/$', UsersDetailView.as_view(), name='users-detail'),
     url(r'^(?P<pk>\d+)/add/$', UsersCreateView.as_view(), name='users-add'),
     url(r'^(?P<pk>\d+)/edit/$', UsersEditView.as_view(), name='users-edit'),
-    url(r'^(?P<pk>\d+)/transfers/$', UsersTransfersView.as_view(), name='users-transfers-report'),
     url(r'^change-password/$', UserPasswordChangeView.as_view(), name='password-change'),
 
 ]
