+++ project_electron/transfer_app/templates/transfer_app/left_sidebar.html
@@ -10,7 +10,7 @@
           </a>
         </li>
         <li>
-          <a href="{% url 'org-transfers' %}">
+          <a href="{% url 'transfers' %}">
             <i class="fa fa-exchange"></i> <span>Transfers</span>
           </a>
         </li>
