+++ project_electron/transfer_app/urls.py
@@ -3,6 +3,7 @@ from orgs.views import *
 from transfer_app.views import *
 
 urlpatterns = [
-    url(r'^$', RecentTransfersView.as_view(), name='org-transfers'),
+    url(r'^$', TransfersView.as_view(), name='transfers'),
+    url(r'^csv/$', TransferDataView.as_view(), name='transfer-data'),
     url(r'^(?P<pk>\d+)$', TransferDetailView.as_view(), name='transfer-detail'),
 ]
